Refactor graphql-client code and file structure
