+++ packages/graphql-client/src/index.ts
@@ -1,2 +1,2 @@
-export { createGraphQLClient } from "./graphql-client";
-export type * from "./types";
+export { createGraphQLClient } from "./graphql-client/graphql-client";
+export type * from "./graphql-client/types";
